[brownfield][android] simplify rn activity in the test app
